Terminate TPU VM in refresh without using backend
